Add div class to table as table-main
Change div class name from table, to table-mane because it was a ...
It was an incorrect decision to assign it to to the same one, as on the
maine page. Changing the background color changing it on the main page
too. I can change the div class name to something else, but eh - it's
working.

Though, It's a good lesson, as if I was working with someone else, that
would be a hard thing to understand.

modified:   recipes/bread.html
modified:   recipes/mead.html
modified:   recipes/roastbeef.html